[cli][pylibs][simulation] allow fractional ping intervals and conform to OT ping defaults
This fixes the OTNS CLI to enable fractional numbers for ping interval, as supported by OT. Also conforms default values to OT's defaults: interval = 1.0, ping size = 8B. Fixes Pylibs API description for 'interval' parameter which was incorrect. Help text for 'ping' fixes to show actual output.

+++ simulation/node.go
@@ -766,8 +766,8 @@ func (node *Node) Reset() {
 	node.Logger.Warn("reset complete")
 }
 
-func (node *Node) Ping(addr string, payloadSize int, count int, interval int, hopLimit int) {
-	cmd := fmt.Sprintf("ping async %s %d %d %d %d", addr, payloadSize, count, interval, hopLimit)
+func (node *Node) Ping(addr string, payloadSize int, count int, interval float64, hopLimit int) {
+	cmd := fmt.Sprintf("ping async %s %d %d %f %d", addr, payloadSize, count, interval, hopLimit)
 	err := node.inputCommand(cmd)
 	if err != nil {
 		node.error(err)
